Include simfile.h from cim.h instead of directly.
Files that need simfile.h defines INCLUDE_SIMFILE_H.
In cim.h we include simfile.h if this define is set
othervice we define __bs file pointers used in the sugnature
of functions to void *.

INCLUDE_SIMFILE_H will never be defined for generated code.

TODO: We could consider to always include simfile.h and other header files
for separately compiled units.  That would require a redesign of
the generated code as these definitions can only be put in the header file.